feat: Checkout Controller functions

+++ src/main/java/com/kashier/interfaces/IInvoice.java
@@ -6,21 +6,21 @@ import java.util.ArrayList;
 
 public interface IInvoice  {
 
-    public float getTotal();
+    public double getTotal();
 
-    public float getDiscount();
+    public double getDiscount();
 
-    public float getFee();
+    public double getFee();
 
     public String getId();
 
     public ArrayList<InvoiceItem> getItems();
 
-    public void setTotal(float total);
+    public void setTotal(double total);
 
-    public void setDiscount(float discount);
+    public void setDiscount(double discount);
 
-    public void setFee(float fee);
+    public void setFee(double fee);
 
     public void setId(String id);
